fix: add command
