+++ ftplugin/calendar.vim
@@ -1,0 +1,3 @@
+set signcolumn=no
+set nonumber
+set norelativenumber
